src/requirements.txt
Flask==1.1.2
Flask-MySQL==1.5.1
mysql-connector-python
arrow